adding files for behavioral testing to ignore, updating cipher key script, updating file-system, server, config and index
