+++ .env.example
@@ -1,14 +1,15 @@
 # DCX configuration
 PORT=3000
-SERVICE_NAME="Decentralized Credential Exchange"
 EXTERNAL_PORT=3000
 EXTERNAL_HOSTNAME=localhost
+SERVICE_NAME="Decentralized Credential Exchange"
 ENCRYPTION_KEY="<to generate enc key: npm run gen-enc-key>"
-ISSUER_ENDPOINT="https://issuer2.example.com/v1/vc"
+ISSUER_SERVICE_ENDPOINT="https://issuer2.example.com/v1/vc"
 TRUSTED_ISSUERS='[{"name": "issuer1", "did": "did:dht:issuer1}]'
-
 # Web5 configuration
 DHT_GATEWAY_ENDPOINT="dht-gateway.example.com"
 DWN_ENDPOINTS="dwn1.example.com,dwn2.example.com"
 DWN_PASSWORD="correct horse battery staple"
-DWN_RECOVERY_PHRASE="correct brown horse jump over large black barrel then hunter shoot robot"
+DWN_RECOVERY_PHRASE="correct brown horse jump over large black barrel then hunt shoot rabbit"
+# DWN_LAST_RECORD_ID_FILE='' # optional
+# DWN_CURSOR_FILE='' # optional
