+++ src/config.ts
@@ -1,18 +1,13 @@
-import { readFile } from "fs/promises";
-
-const DWN_LAST_RECORD_ID_FILE = `${process.cwd()}/lastRecordId`;
-const DWN_CURSOR_FILE = `${process.cwd()}/cursor.json`;
-
-const DWN_LAST_RECORD_ID = await readFile(DWN_LAST_RECORD_ID_FILE) || '';
-const DWN_CURSOR = await readFile(DWN_CURSOR_FILE) || '';
-
+import { config } from 'dotenv';
+config();
 export class Config {
-  public DWN_LAST_RECORD_ID_FILE = DWN_LAST_RECORD_ID_FILE;
-  public DWN_CURSOR_FILE = DWN_CURSOR_FILE;
-  public DWN_LAST_RECORD_ID = DWN_LAST_RECORD_ID;
-  public DWN_CURSOR = DWN_CURSOR;
+  public DWN_LAST_RECORD_ID_FILE = process.env.DWN_LAST_RECORD_ID_FILE || `${process.cwd()}/lastRecordId`;
+  public DWN_CURSOR_FILE = process.env.DWN_CURSOR_FILE || `${process.cwd()}/cursor.json`;
+
   public ISSUER_SERVICE_ENDPOINT = process.env.ISSUER_SERVICE_ENDPOINT || '';
-  public DHT_GATEWAY_ENDPOINT = process.env.DHT_GATEWAY_ENDPOINT || '';
+  public TRUSTED_ISSUERS = (process.env.TRUSTED_ISSUERS && JSON.parse(process.env.TRUSTED_ISSUERS)) || [{ "name": "mx", "did": "did:dht:sa713dw7jyg44ejwcdf8iqcseh7jcz51wj6fjxbooj41ipeg76eo" }];
+
+  public DHT_GATEWAY_ENDPOINT = process.env.DHT_GATEWAY_ENDPOINT || '0.0.0.0:8305';
   public DWN_ENDPOINTS = process.env.DWN_ENDPOINTS?.split(',') || [];
 
   private _CIPHER_KEY: string;
@@ -39,3 +34,12 @@ export class Config {
 }
 
 export const dcxConfig = new Config();
+
+console.log("dcxConfig", dcxConfig)
+
+/*
+  TODO: work these async loading function into config class
+  import { readFileToJSON, readFileToString } from './utils/file-system.js';
+  public DWN_LAST_RECORD_ID = readFileToString(this.DWN_LAST_RECORD_ID_FILE).then(lastRecordId => lastRecordId).catch(console.error) ?? '';
+    public DWN_CURSOR = readFileToJSON(this.DWN_CURSOR_FILE).then(cursor => cursor).catch(console.error) ?? {};
+*/
